1. fix bug 10: "  RAG uses £ GBP pricing; users/listings use $ USD "
prompt the supevisor to not give price and year justifications on the "why this model" field as it should be general to the make+model level not a listing spesific info.
this also eliminate bug 10 because no price are talken -> no £ as well

2. fixed simple bug: the years was presented as float in the response.

+++ agents/search_agent/listings_retriever.py
@@ -51,10 +51,11 @@ class ListingsRetriever:
             if col in self.df.columns:
                 self.df[col] = self.df[col].fillna("").astype(str).str.lower().str.strip()
         
-        # Convert numeric columns
+        # Convert numeric columns — use nullable Int64 so NaN rows don't
+        # force the whole column to float (which would serialize as 2016.0 in JSON)
         for col in ("price", "year"):
             if col in self.df.columns:
-                self.df[col] = pd.to_numeric(self.df[col], errors="coerce")
+                self.df[col] = pd.to_numeric(self.df[col], errors="coerce").astype("Int64")
         
         print(f"Loaded {len(self.df)} listings from {self.csv_path}")
